Create README.md
Photo app that allows the user to browse and see Flickr public contents.

The code is written with Kotlin and follows android recommended guidelines. It particularly makes the best use of the android jetpack libraries:

Room is used to keep the user search queries
Workmanager to check if there is new content available. If the app is not on the foreground a notification is sent to the user.
pagination for partial network data loading
databinding which simplifies the code greatly
To communicate with Flickr servers, Retrofit library has been used. Also, the images are downaloed and processed with Glide.